combined example
fused the standard pitch extraction examples to a single example, where
the used pitch detection algorithm is given as an input argument
